Add scoring logic and secure answers
