+++ server/api/questions/junior1/read.get.ts
@@ -1,10 +1,8 @@
 import { promises as fs } from 'node:fs'
-import {join} from 'node:path'
+import { join } from 'node:path'
 
-// TODO: Кажется, это можно удалить
-export default defineEventHandler(async() => {
-  const file = join(process.cwd(),
-    'data', 'junior1.json')
-    const json = await fs.readFile(file, 'utf-8')
-    return JSON.parse(json)
+export default defineEventHandler(async () => {
+  const file = join(process.cwd(), 'data', 'junior1.json')
+  const json = await fs.readFile(file, 'utf-8')
+  return JSON.parse(json)
 })
