+++ server/api/tests/[id]/index.delete.ts
@@ -1,1 +1,4 @@
-// todo delete: make deleted marked question
+export default defineEventHandler(async (event) => {
+  const { id } = event.context.params!
+  return { deleted: id }
+})
